fix some auto-format yuckiness

+++ betamax-core/src/test/groovy/co/freeside/betamax/recorder/TapeModeSpec.groovy
@@ -29,15 +29,14 @@ import static java.net.HttpURLConnection.HTTP_OK
 @Unroll
 class TapeModeSpec extends Specification {
 
-    @Shared @AutoCleanup('deleteDir') File tapeRoot = Files.createTempDir()
-
+    @Shared @AutoCleanup("deleteDir") File tapeRoot = Files.createTempDir()
     @Shared Recorder recorder = new Recorder(tapeRoot: tapeRoot)
 
-    @Shared @AutoCleanup('stop') SimpleServer endpoint = new SimpleServer(HelloHandler)
+    @Shared @AutoCleanup("stop") SimpleServer endpoint = new SimpleServer(HelloHandler)
 
     HttpHandler handler = new DefaultHandlerChain(recorder)
 
-    Request request = new BasicRequest('GET', endpoint.url)
+    Request request = new BasicRequest("GET", endpoint.url)
 
     void setupSpec() {
         endpoint.start()
@@ -47,37 +46,31 @@ class TapeModeSpec extends Specification {
         recorder.ejectTape()
     }
 
-    void 'in #mode mode the proxy rejects a request if no recorded interaction exists'() {
-        given:
-        'a read-only tape is inserted'
-        recorder.insertTape('read only tape', [mode: mode])
+    void "in #mode mode the proxy rejects a request if no recorded interaction exists"() {
+        given: "a read-only tape is inserted"
+        recorder.insertTape("read only tape", [mode: mode])
 
-        when:
-        'a request is made that does not match anything recorded on the tape'
+        when: "a request is made that does not match anything recorded on the tape"
         handler.handle(request)
 
-        then:
-        'the proxy rejects the request'
+        then: "the proxy rejects the request"
         thrown NonWritableTapeException
 
         where:
         mode << [READ_ONLY, READ_SEQUENTIAL]
-        request = new BasicRequest('GET', endpoint.url)
+        request = new BasicRequest("GET", endpoint.url)
     }
 
-    void 'in #mode mode a new interaction recorded'() {
-        given:
-        'an empty write-only tape is inserted'
-        new File(tapeRoot, 'blank_tape_' + mode + '.yaml').delete()
-        recorder.insertTape('blank tape ' + mode, [mode: mode])
+    void "in #mode mode a new interaction is recorded"() {
+        given: "an empty write-only tape is inserted"
+        new File(tapeRoot, "blank_tape_" + mode + ".yaml").delete()
+        recorder.insertTape("blank tape " + mode, [mode: mode])
         def tape = recorder.tape
 
-        when:
-        'a request is made'
+        when: "a request is made"
         handler.handle(request)
 
-        then:
-        'the interaction is recorded'
+        then: "the interaction is recorded"
         tape.size() == old(tape.size()) + 1
 
         cleanup:
@@ -87,10 +80,9 @@ class TapeModeSpec extends Specification {
         mode << [READ_WRITE, WRITE_ONLY, WRITE_SEQUENTIAL]
     }
 
-    void 'in write-only mode the proxy overwrites an existing matching interaction'() {
-        given:
-        'an existing tape file is inserted in write-only mode'
-        def tapeFile = new File(tapeRoot, 'write_only_tape.yaml')
+    void "in write-only mode the proxy overwrites an existing matching interaction"() {
+        given: "an existing tape file is inserted in write-only mode"
+        def tapeFile = new File(tapeRoot, "write_only_tape.yaml")
         tapeFile.text = """\
 !tape
 name: write only tape
@@ -105,26 +97,23 @@ interactions:
     headers: {}
     body: Previous response made when endpoint was down.
 """
-        recorder.insertTape('write only tape', [mode: WRITE_ONLY])
+        recorder.insertTape("write only tape", [mode: WRITE_ONLY])
         def tape = recorder.tape
 
-        when:
-        'a request is made that matches a request already recorded on the tape'
+        when: "a request is made that matches a request already recorded on the tape"
         handler.handle(request)
 
-        then:
-        'the previously recorded request is overwritten'
+        then: "the previously recorded request is overwritten"
         tape.size() == old(tape.size())
         tape.interactions[-1].response.status == HTTP_OK
         tape.interactions[-1].response.body
     }
 
-    @Issue('https://github.com/robfletcher/betamax/issues/7')
-    @Issue('https://github.com/robfletcher/betamax/pull/70')
-    void 'in write-sequential mode the proxy records additional interactions'() {
-        given:
-        'an existing tape file is inserted in write-sequential mode'
-        def tapeFile = new File(tapeRoot, 'write_sequential_tape.yaml')
+    @Issue("https://github.com/robfletcher/betamax/issues/7")
+    @Issue("https://github.com/robfletcher/betamax/pull/70")
+    void "in write-sequential mode the proxy records additional interactions"() {
+        given: "an existing tape file is inserted in write-sequential mode"
+        def tapeFile = new File(tapeRoot, "write_sequential_tape.yaml")
         tapeFile.text = """\
 !tape
 name: write sequential tape
@@ -139,15 +128,13 @@ interactions:
     headers: {}
     body: Previous response made when endpoint was down.
 """
-        recorder.insertTape('write sequential tape', [mode: WRITE_SEQUENTIAL])
+        recorder.insertTape("write sequential tape", [mode: WRITE_SEQUENTIAL])
         def tape = recorder.tape
 
-        when:
-        'a request is made that matches a request already recorded on the tape'
+        when: "a request is made that matches a request already recorded on the tape"
         handler.handle(request)
 
-        then:
-        'the previously recorded request is overwritten'
+        then: "the previously recorded request is overwritten"
         tape.size() == old(tape.size()) + 1
         tape.interactions[-1].response.status == HTTP_OK
         tape.interactions[-1].response.body
